2.0.OSubMarin - Moved the TechType and CraftTreeType prepatch checks out of EnumCacheManager and into their respective patcher classes - Remove some logging statements and commented code -
