+++ SMLHelper/Patchers/CraftTreeTypePatcher.cs
@@ -18,9 +18,7 @@
             new EnumCacheManager<CraftTree.Type>(
                 enumTypeName: CraftTreeTypeEnumName,
                 startingIndex: startingIndex,
-                bannedIndices: ExtBannedIdManager.GetBannedIdsFor(CraftTreeTypeEnumName));
-        
-        #region Adding CraftTreeTypes and TreeRoots
+                bannedIndices: ExtBannedIdManager.GetBannedIdsFor(CraftTreeTypeEnumName, PreRegisteredCraftTreeTypes()));
 
         internal static ModCraftTreeRoot CreateCustomCraftTreeAndType(string name, out CraftTree.Type craftTreeType)
         {
@@ -53,7 +51,37 @@
             return customTreeRoot;
         }
 
-        #endregion
+        private static List<int> PreRegisteredCraftTreeTypes()
+        {
+            // Make sure to exclude already registered CraftTreeTypes.
+            // Be aware that this approach is still subject to race conditions.
+            // Any mod that patches after this one will not be picked up by this method.
+            // For those cases, there are additional ways of excluding these IDs.
+
+            var bannedIndices = new List<int>();
+
+            var enumValues = Enum.GetValues(typeof(CraftTree.Type));
+            foreach (var enumValue in enumValues)
+            {
+                if (enumValue != null)
+                {
+                    int realEnumValue = (int)enumValue;
+
+                    if (realEnumValue > startingIndex)
+                    {
+                        if (!bannedIndices.Contains(realEnumValue))
+                        {
+                            bannedIndices.Add(realEnumValue);
+                        }
+                    }
+
+                }
+            }
+
+            Logger.Log($"Finished known CraftTreeType exclusion. {bannedIndices.Count} IDs were added in ban list.");
+
+            return bannedIndices;
+        }
 
         #region Patches
